moving custom components to the correct folder
