+++ src/components/AddProductModal.tsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from 'react';
-import Modal from '@/components/ui/modal';
+import Modal from '@/components/modal';
 import { Button } from '@/components/ui/button';
 
 interface AddProductModalProps {
